move output codes declarations
