+++ wingetui/PackageManagers/PackageClasses.py
@@ -203,3 +203,23 @@ class PackageManagerModule():
         """
         Force update package manager's sources
         """
+
+RETURNCODE_OPERATION_SUCCEEDED = 0
+RETURNCODE_NO_APPLICABLE_UPDATE_FOUND = 92849
+RETURNCODE_NEEDS_RESTART = 3
+
+LIST_RETURNCODES_OPERATION_SUCCEEDED = (RETURNCODE_OPERATION_SUCCEEDED, RETURNCODE_NO_APPLICABLE_UPDATE_FOUND, RETURNCODE_NEEDS_RESTART)
+
+RETURNCODE_INCORRECT_HASH = 2
+RETURNCODE_NEEDS_ELEVATION = 1603
+RETURNCODE_NEEDS_SCOOP_ELEVATION = 1602
+
+ENABLE_WINGETUI_NOTIFICATIONS = True
+ENABLE_SUCCESS_NOTIFICATIONS = True
+ENABLE_ERROR_NOTIFICATIONS = True
+ENABLE_UPDATES_NOTIFICATIONS = True
+
+class BlacklistMethod():
+    Legacy = 0
+    SpecificVersion = 1
+    AllVersions = 2
